% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autorefresh.R
\name{autorefresh}
\alias{autorefresh}
\title{autorefresh}
\usage{
autorefresh()
}
\description{
Starts a Shiny app that tracks changes in the {projectroot}/R folder and updates the interface of the app accordingly.
The function needs to be a run from a rstudio Job or terminal session for it to be able to do its work.
}
\seealso{
refresh
}
