% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh}
\alias{refresh}
\title{autoreload for shiny app that contains multiple files (outside of app.R)}
\usage{
refresh()
}
\arguments{
\item{app.R}{The location of your app.R file. It is assumed to be at the root of your project.}
}
\description{
An addin is added to Rstudio which can be binded to a keyboard shortcut.
Alternatively, a Shiny app can be launched which monitors file changes and updates the modified date & time of "app.R".
This results in an automatic refresh of your app.
}
\seealso{
autorefresh
}
